#250 Moved the Particle class to utility.
