+++ atomic/tools.py
@@ -1,92 +1,8 @@
 from utility.constants import Constants
 from utility.exceptions import InputError
-#from atomic.coefficients import ATOMIC_SOURCES
 import numpy as np
 import h5py
 import scipy.constants as sc
-#from renate_od.atomic import cross_section
-
-CONST = Constants()
-
-
-class Particle(object):
-    def __init__(self, label='', charge=0, mass_number=0, atomic_number=0, mass=None):
-        if isinstance(label, str):
-            self.label = label
-        else:
-            raise InputError('Label is expected to be of string type.')
-        try:
-            atomic_number = int(atomic_number)
-            mass_number = int(mass_number)
-            charge = int(charge)
-        except:
-            raise InputError('The atomic number, mass number and charge are expected to be integers.')
-        self.atomic_number = atomic_number
-        self.charge = charge
-        if (mass_number - self.atomic_number) >= 0:
-            self.mass_number = mass_number
-            self.neutron_number = self.mass_number - self.atomic_number
-        if mass is None:
-            if tuple(self) == (-1, 0, 0):
-                self.mass = CONST.electron_mass
-            else:
-                self.mass = self.neutron_number*CONST.neutron_mass +\
-                    self.atomic_number*CONST.proton_mass
-        else:
-            self.mass = mass
-
-    def update_mass(self, mass):
-        self.mass = mass
-
-    def update_atomic_number(self, atomic_number):
-        self.atomic_number = atomic_number
-
-    def update_mass_number(self, mass_number):
-        self.mass_number = mass_number
-        self.neutron_number = self.mass_number - self.atomic_number
-
-    def update_charge(self, charge):
-        self.charge = charge
-
-    def __str__(self):
-        return str(self.label)
-
-    def __iter__(self):
-        for x in [self.charge, self.atomic_number, self.mass_number]:
-            yield x
-
-    def __repr__(self):
-        return 'Particle: ' + str(self.label) + '\t (q,Z,A) =  (' + str(self.charge) + ',' + str(self.atomic_number) + \
-               ',' + str(self.mass_number) + ')'
-
-
-class Transition(object):
-    def __init__(self, projectile=Particle, target=Particle, from_level=str, to_level=None, trans=str):
-        self.projectile = projectile
-        self.target = target
-        if isinstance(from_level, str) and isinstance(trans, str):
-            self.from_level = from_level
-            if trans in ['ex', 'de-ex', 'eloss', 'ion', 'cx']:
-                self.name = trans
-            else:
-                InputError('The provided transition is not supported. '
-                           'Supported transitions are: ex, de-ex, eloss, ion and cx.')
-            if (to_level is None) or isinstance(to_level, str):
-                self.to_level = to_level
-            else:
-                InputError('The provided end state for the electron transition is not valid. Str or None is expected.')
-        else:
-            InputError('Expected input data format for <from_level>, <to_level> and <trans> to be of str type.')
-
-    def __str__(self):
-        if self.name in ['cx', 'eloss', 'ion']:
-            return self.from_level + '-' + self.name
-        else:
-            return self.from_level + '-' + self.to_level
-
-    def __repr__(self):
-        return 'Collision of: '+str(self.projectile)+' + '+str(self.target)+' with transition: ' + \
-               self.name+' | from level: '+self.from_level+' | to_level: '+str(self.to_level)
 
 
 class RENATE_H_hdf_generator():
