+++ utility/__init__.py
@@ -1,3 +1,5 @@
 from utility import getdata
 from utility import convert
 from utility import exceptions
+from utility.particle import Particle
+from utility.transition import Transition
